New Levels + Game
If you navigate in the unity, builds, windows, 86, 0.3.0, raw. you will find the game files for the new version, I will try to get the installer shortly, you will need the whole raw folder to be installed for the game to run. the difficulty curve is dramatic from no difficulty to really difficult. There are some levels missing in between as I’m running low on time rn